Default without needing to manage state
