+++ playbook/app/pb_kits/playbook/pb_advanced_table/docs/_advanced_table_column_visibility.md
@@ -1,0 +1,4 @@
+The `columnVisibilityControl` prop allows users to toggle the visibility of table columns dynamically.
+
+The default can be enabled simply by passing `{ default:true }` to the prop as shown. This will render the header with the icon enabled dropdown. The dropdown contains all columns present in the Table and any can be toggled on or off via the checkboxes. 
+
