+++ playbook/app/pb_kits/playbook/pb_advanced_table/docs/_advanced_table_column_visibility_multi.md
@@ -1,0 +1,1 @@
+The `columnVisibilityControl` prop can also be used with multi-header columns as shown.
